Add communication channel from Hap to graphics mode.
When running in graphics mode, Hap can now add graphics commands to a
channel ('TChan') in its environment, where each command is an 'IO'
action that takes an 'SDL.Renderer'. The main thread continually polls
this channel in the graphics loop and executes actions accordingly. When
running in text output mode, no channel is created, and graphics
intrinsics are no-ops. Those new intrinsics are:

 * graphics_background_set(r: int, g: int, b: int): null

   Sets the background color to an RGB8 value.

 * graphics_clear(): null

   Clears the graphics window to the current background color.

 * graphics_present(): null

   Presents the render buffer on the screen.